@@ -1,29 +1,29 @@
 using NUnit.Framework;
+using NUnit.Framework.Legacy;
 using Wikiled.Console.Arguments;
 
-namespace Wikiled.Console.Tests.Arguments
+namespace Wikiled.Console.Tests.Arguments;
+
+[TestFixture]
+public class CommandLineDictionaryTests
 {
-    [TestFixture]
-    public class CommandLineDictionaryTests
+    private CommandLineDictionary instance;
+
+    [SetUp]
+    public void SetUp()
     {
-        private CommandLineDictionary instance;
+        instance = CreateCommandLineDictionary();
 
-        [SetUp]
-        public void SetUp()
-        {
-            instance = CreateCommandLineDictionary();
+    }
 
-        }
+    [Test]
+    public void Construct()
+    {
+        ClassicAssert.IsNotNull(instance);
+    }
 
-        [Test]
-        public void Construct()
-        {
-            Assert.IsNotNull(instance);
-        }
-
-        private CommandLineDictionary CreateCommandLineDictionary()
-        {
-            return new CommandLineDictionary();
-        }
+    private CommandLineDictionary CreateCommandLineDictionary()
+    {
+        return new CommandLineDictionary();
     }
 }
