@@ -1,17 +1,16 @@
 ﻿using System;
 
-namespace Wikiled.Console.Arguments
+namespace Wikiled.Console.Arguments;
+
+public class ExecutionContext
 {
-    public class ExecutionContext
+    public ExecutionContext(string name, AppConfig config)
     {
-        public ExecutionContext(string name, AppConfig config)
-        {
-            CommandName = name ?? throw new ArgumentNullException(nameof(name));
-            Config = config ?? throw new ArgumentNullException(nameof(config));
-        }
-
-        public string CommandName { get; }
-
-        public AppConfig Config { get; }
+        CommandName = name ?? throw new ArgumentNullException(nameof(name));
+        Config = config ?? throw new ArgumentNullException(nameof(config));
     }
+
+    public string CommandName { get; }
+
+    public AppConfig Config { get; }
 }
