@@ -1,11 +1,10 @@
-﻿namespace Wikiled.Console.Arguments
-{
-    public interface IAutoStarter
-    {
-        string Name { get; }
+﻿namespace Wikiled.Console.Arguments;
 
-        IAutoStarter RegisterCommand<T, TConfig>(string name)
-            where T : ICommand
-            where TConfig : ICommandConfig, new();
-    }
+public interface IAutoStarter
+{
+    string Name { get; }
+
+    IAutoStarter RegisterCommand<T, TConfig>(string name)
+        where T : ICommand
+        where TConfig : ICommandConfig, new();
 }
